# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PROOF API
  description: PROOF 서비스 REST API 명세서
  version: 1.0.0
  contact:
    name: PROOF Team
servers:
  - url: http://localhost:8000
    description: 로컬 개발 서버
...
paths:
...
  /api/reflections:
    get:
      tags:
        - Reflections
      summary: 회고 목록 조회
      description: 사용자의 회고 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: space_id
          in: query
          description: 스페이스 ID 필터
          schema:
            type: string
            format: uuid
        - name: type
          in: query
          description: 회고 유형 필터
          schema:
            type: string
            enum: [chatbot, template, survey]
        - name: limit
          in: query
          description: 조회 개수
          schema:
            type: integer
            default: 20
        - name: offset
          in: query
          description: 오프셋
          schema:
            type: integer
            default: 0
      responses:
        '200':
          description: 회고 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                    example: 50
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Reflection'

    post:
      tags:
        - Reflections
      summary: 회고 생성
      description: 새로운 회고를 생성합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - type
                - content
              properties:
                space_id:
                  type: string
                  format: uuid
                  description: 스페이스 ID (선택)
                template_id:
                  type: string
                  format: uuid
                  description: 템플릿 ID (선택)
                type:
                  type: string
                  enum: [chatbot, template, survey]
                  description: 회고 유형
                  example: "chatbot"
                title:
                  type: string
                  description: 제목
                  example: "1주차 회고"
                content:
                  type: object
                  description: 회고 내용 (STAR, Q&A 등)
                  example:
                    situation: "팀 프로젝트에서 마케팅 전략 수립"
                    task: "경쟁사 분석 및 포지셔닝 전략 개발"
                    action: "시장 조사 및 SWOT 분석 수행"
                    result: "차별화된 마케팅 전략 도출"
                mood_before:
                  type: string
                  enum: [good, neutral, bad]
                  example: "neutral"
                mood_after:
                  type: string
                  enum: [good, neutral, bad]
                  example: "good"
                tags:
                  type: array
                  items:
                    type: string
                  example: ["마케팅", "전략", "분석"]
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reflection'
  /api/reflections/{reflection_id}:
    get:
      tags:
        - Reflections
      summary: 회고 상세 조회
      description: 특정 회고의 상세 정보를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: reflection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: 회고 상세 정보
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reflection'

    patch:
      tags:
        - Reflections
      summary: 회고 수정
      description: 회고 내용을 수정합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: reflection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                content:
                  type: object
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reflection'

    delete:
      tags:
        - Reflections
      summary: 회고 삭제
      description: 회고를 삭제합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: reflection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: 삭제 성공
  /api/reflections/{reflection_id}/ai-feedback:
    post:
      tags:
        - Reflections
      summary: AI 피드백 생성
      description: 회고에 대한 AI 피드백을 생성합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: reflection_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI 피드백 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  ai_feedback:
                    type: string
                    example: "이번 활동에서 체계적인 분석 능력을 잘 발휘하셨네요..."
                  progress_score:
                    type: integer
                    example: 85
  /api/reflections/stats:
    get:
      tags:
        - Reflections
      summary: 회고 통계 조회
      description: 사용자의 회고 통계를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: period
          in: query
          description: 기간
          schema:
            type: string
            enum: [week, month, quarter]
            default: week
      responses:
        '200':
          description: 회고 통계
          content:
            application/json:
              schema:
                type: object
                properties:
                  total_reflections:
                    type: integer
                    example: 24
                  consecutive_days:
                    type: integer
                    example: 7
                  active_spaces:
                    type: integer
                    example: 3
                  growth_keywords:
                    type: array
                    items:
                      type: string
                    example: ["협업", "분석", "기획"]
                  mood_summary:
                    type: object
                    properties:
                      good:
                        type: integer
                      neutral:
                        type: integer
                      bad:
                        type: integer
  /api/reflections/story:
    get:
      tags:
        - Reflections
      summary: 성장 스토리 조회
      description: 기간별 성장 스토리를 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: period
          in: query
          required: true
          description: 조회 기간
          schema:
            type: string
            enum: [week, month, quarter]
      responses:
        '200':
          description: 성장 스토리
          content:
            application/json:
              schema:
                type: object
                properties:
                  period:
                    type: string
                    example: "week"
                  summary:
                    type: string
                    example: "이번 주에는 3개의 활동을 완료하셨습니다."
                  activities:
                    type: array
                    items:
                      type: object
                      properties:
                        type:
                          type: string
                        count:
                          type: integer
                  positive_moments:
                    type: array
                    items:
                      type: string
                  negative_moments:
                    type: array
                    items:
                      type: string
                  strengths:
                    type: array
                    items:
                      type: string
                  recommendations:
                    type: array
                    items:
                      type: object

  # ==================== Micro Logs ====================
  /api/reflections/micro:
    get:
      tags:
        - Micro Logs
      summary: 마이크로 로그 목록 조회
      description: 사용자의 마이크로 로그 목록을 조회합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: 마이크로 로그 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/MicroLog'

    post:
      tags:
        - Micro Logs
      summary: 마이크로 로그 생성
      description: 간단한 활동 기록을 생성합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - activity_type
                - mood
              properties:
                activity_type:
                  type: string
                  enum: [공모전, 프로젝트, 동아리, 인턴, 대외활동, 기타]
                  example: "프로젝트"
                memo:
                  type: string
                  example: "팀 미팅 진행 및 역할 분담"
                mood:
                  type: string
                  enum: [good, neutral, bad]
                  example: "good"
                mood_reason:
                  type: string
                  example: "팀원들과 소통이 잘 되었음"
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MicroLog'
  /api/reflections/micro/{log_id}/tags:
    post:
      tags:
        - Micro Logs
      summary: AI 태그 생성
      description: 마이크로 로그에 AI 태그를 생성합니다
      parameters:
        - $ref: '#/components/parameters/UserIdHeader'
        - name: log_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: AI 태그 생성 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  tags:
                    type: array
                    items:
                      type: string
                    example: ["협업", "소통", "리더십"]

  # ==================== Templates ====================
...
components:
  parameters:
    UserIdHeader:
      name: x-user-id
      in: header
      required: true
      description: 사용자 ID
      schema:
        type: string
        example: "john_doe"
  schemas:
...
    Reflection:
      type: object
      properties:
        reflection_id:
          type: string
          format: uuid
        user_id:
          type: string
        space_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        type:
          type: string
          enum: [chatbot, template, survey]
        title:
          type: string
        content:
          type: object
        ai_feedback:
          type: string
        progress_score:
          type: integer
        mood_before:
          type: string
          enum: [good, neutral, bad]
        mood_after:
          type: string
          enum: [good, neutral, bad]
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    MicroLog:
      type: object
      properties:
        log_id:
          type: string
          format: uuid
        user_id:
          type: string
        activity_type:
          type: string
          enum: [공모전, 프로젝트, 동아리, 인턴, 대외활동, 기타]
        memo:
          type: string
        mood:
          type: string
          enum: [good, neutral, bad]
        mood_reason:
          type: string
        tags:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
...